+++ src/daq_cli/cli/app.py
@@ -6,6 +6,7 @@ from daq_cli.cli.decode import app as decode_app
 from daq_cli.cli.group import app as group_app
 from daq_cli.cli.monitor import app as monitor_app
 from daq_cli.cli.profile import app as profile_app
+from daq_cli.cli.tcm import app as tcm_app
 from daq_cli.cli.wave import app as wave_app
 
 app = typer.Typer(
@@ -19,4 +20,5 @@ app.add_typer(group_app, name="group")
 app.add_typer(acquire_app, name="acquire")
 app.add_typer(decode_app, name="decode")
 app.add_typer(monitor_app, name="monitor")
+app.add_typer(tcm_app, name="tcm")
 app.add_typer(wave_app, name="wave")